+++ RELEASES.md
@@ -1,6 +1,10 @@
 Version 0.10.0 (UNRELEASED)
 ===========================
 
+New Features
+------------
+* [FIXME: ELABORATE]: Vector types.
+
 
 Breaking Changes
 ----------------
@@ -8,6 +12,7 @@ Breaking Changes
   `Event` or `EventList` argument when setting the wait list using `::ewait`
   or `::ewait_opt`. Awkward type ascriptions when not passing a wait list can
   now be removed due to the use of a trait object argument type.
+* [FIXME: ELABORATE]: Fill now takes a vector type.